Adding rating mechanism to adjust for leagues promotion/relegation.
 - Each network has a data structure with the leagues and seasons
playing teams.
 - If a team is added to a season where he was not before, the mean of
the previous teams is added.
 - However, this creates an issue in the case of intermediate leagues:
The second division receives two new teams from the first division while
loosing four teams (2 promoted and 2 relegated to the third).
 - By now, the approach is that teams relegating to the third (i.e., not
having a playing structure in the current season as we are only covering
up to the second) are not used to computed the entry rating.